Refactor of KHeader and KDat: Limit Scope
We limit our scope to some of the more important header fields
We enforce that these are set with best effort
and dump everyhting else in misc.

If you load in data where one of the "important" fields is missing,
you have to provide default values in a dictionary.

Encoding Limits have defaults.
Acceeleraton factors is commented our for now.